requirements.txt
flask
Flask-SQLAlchemy
pyjwt
datetime
uuid
python-dotenv
flask-cors
mysqlclient